fix: Enhance folder removal logic in Remove-Lab function to check if path exists

+++ AutomatedLabCore/functions/Core/Remove-Lab.ps1
@@ -161,7 +161,7 @@
                 }
 
                 #Only remove folder for VMs if folder is empty
-                if ($Script:data.Target.Path -and (-not (Get-ChildItem -Path $Script:data.Target.Path)))
+                if ($Script:data.Target.Path -and (Test-Path -Path $Script:data.Target.Path) -and (-not (Get-ChildItem -Path $Script:data.Target.Path)))
                 {
                     Remove-Item -Path $Script:data.Target.Path -Recurse -Force -Confirm:$false
                 }
